refactor: viewConstant에서 exception으로 패키지 이름 변경
