+++ Assignment/bank/src/main/java/sopt/bank/view/InputView.java
@@ -3,7 +3,7 @@ package sopt.bank.view;
 import sopt.bank.view.dto.DepositDTO;
 import sopt.bank.view.dto.TransferDTO;
 import sopt.bank.view.dto.WithdrawDTO;
-import sopt.bank.view.viewConstant.IllegalArgumentExceptionType;
+import sopt.bank.view.exception.IllegalArgumentExceptionType;
 
 import java.util.Scanner;
